backend/requirements.txt
# Flask e extensões
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Banco de dados
mysql-connector-python==8.2.0

# Segurança
bcrypt==4.1.2
Werkzeug==3.0.1

# Utilitários
python-dotenv==1.0.0
python-dateutil==2.8.2

# Servidor WSGI para produção
gunicorn==21.2.0

# Desenvolvimento (opcional)
pytest==7.4.3
pytest-flask==1.3.0

